SERVICE:FIXED

+++ yiyingliService/src/cn/yiyingli/Dao/Impl/RewardDaoImpl.java
@@ -1,7 +1,6 @@
 package cn.yiyingli.Dao.Impl;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
@@ -34,23 +33,25 @@ public class RewardDaoImpl extends HibernateDaoSupport implements RewardDao {
 		session.flush();
 		Query query = session.createSQLQuery(
 				"update teacher set teacher.REWARDNUMBER=(select count(*) from reward where reward.TEACHER_ID='"
-						+ reward.getTeacherId() + "' and reward.FINISHPAY=true) where teacher.TEACHER_ID=" + reward.getTeacherId());
+						+ reward.getTeacherId() + "' and reward.FINISHPAY=true) where teacher.TEACHER_ID="
+						+ reward.getTeacherId());
 		query.executeUpdate();
 		if (reward.getPassageId() != null) {
 			query = session.createSQLQuery(
 					"update passage set passage.REWARDNUMBER=(select count(*) from reward where reward.PASSAGE_ID='"
-							+ reward.getPassageId() + "' and reward.FINISHPAY=true) where passage.PASSAGE_ID=" + reward.getPassageId());
+							+ reward.getPassageId() + "' and reward.FINISHPAY=true) where passage.PASSAGE_ID="
+							+ reward.getPassageId());
 			query.executeUpdate();
 		}
 	}
 
 	@Override
 	public Long queryRewardNoByTeacher(long teacherId) {
-		String hql = "select count(*) from Reward r where r.teacherId=" + teacherId+" and r.finishPay=true";
+		String hql = "select count(*) from Reward r where r.teacherId=" + teacherId + " and r.finishPay=true";
 		Session session = getSessionFactory().getCurrentSession();
 		Query query = session.createQuery(hql);
-		BigInteger peopleNo = (BigInteger) query.uniqueResult();
-		return peopleNo.longValue();
+		long peopleNo = (long) query.uniqueResult();
+		return peopleNo;
 	}
 
 	@Override
@@ -115,7 +116,8 @@ public class RewardDaoImpl extends HibernateDaoSupport implements RewardDao {
 		list = getHibernateTemplate().executeFind(new HibernateCallback<List<Reward>>() {
 			@Override
 			public List<Reward> doInHibernate(Session session) throws HibernateException, SQLException {
-				String hql = "from Reward r where teacherId=" + teacherId + " and r.finishPay=true ORDER BY r.createTime DESC";
+				String hql = "from Reward r where teacherId=" + teacherId
+						+ " and r.finishPay=true ORDER BY r.createTime DESC";
 				Query query = session.createQuery(hql);
 				query.setFirstResult((page - 1) * pageSize);
 				query.setMaxResults(pageSize);
